requirements.txt
# IoT Security Monitoring System - Python Dependencies

# MQTT Communication
paho-mqtt>=1.6.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0,<2.0.0  # Compatible with most packages

# AI/Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Dashboard & Visualization
streamlit>=1.28.0
plotly>=5.17.0

# Configuration Management
PyYAML>=6.0

# Database
# SQLite is built-in to Python

# Serial Communication (for Arduino)
pyserial>=3.5

# Utilities
python-dateutil>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Notifications
# smtplib is built-in
# python-telegram-bot==20.7  # Uncomment if using Telegram alerts
